Adjustments to the API fields of the activities

+++ app/Http/Controllers/EventsController.php
@@ -133,7 +133,17 @@ class EventsController extends Controller
     {
         //
         $event = Activity::find($id);
-        $event->delete();
-        return redirect()->route('events.index')->with('success', 'Activity deleted successfully');
+        if ($event) {
+
+            $imagePath = public_path('imgs/' . $event->image);
+            if ($event->image !== 'default.png' && file_exists($imagePath)) {
+                unlink($imagePath);
+            }
+            $event->delete();
+    
+            return redirect()->route('events.index')->with('success', 'Activity deleted successfully');
+        } else {
+            return redirect()->route('events.index')->with('error', 'Activity not found');
+        }
     }
 }
